+++ desktop/packages/mullvad-vpn/src/renderer/components/views/select-location/components/location-lists/hooks/index.ts
@@ -2,3 +2,4 @@ export * from './use-handle-select-entry-location';
 export * from './use-handle-select-exit-location';
 export * from './use-has-visible-locations';
 export * from './use-has-searched';
+export * from './use-has-custom-lists';
