+++ desktop/packages/mullvad-vpn/src/renderer/components/views/select-location/components/location-lists/hooks/use-has-custom-lists.ts
@@ -1,0 +1,9 @@
+import { useLocationListsContext } from '../LocationListsContext';
+
+export function useHasCustomLists() {
+  const { customListLocations } = useLocationListsContext();
+
+  const hasCustomLists = customListLocations.length > 0;
+
+  return hasCustomLists;
+}
