+++ desktop/packages/mullvad-vpn/src/renderer/components/views/select-location/components/location-lists/hooks/use-has-visible-locations.ts
@@ -1,7 +1,7 @@
-import { useSelectLocationViewContext } from '../../../SelectLocationViewContext';
+import { useLocationListsContext } from '../LocationListsContext';
 
 export function useHasSearchedLocations() {
-  const { countryLocations } = useSelectLocationViewContext();
+  const { countryLocations } = useLocationListsContext();
 
   const hasSearchedLocations = countryLocations.length > 0;
