+++ desktop/packages/mullvad-vpn/src/renderer/components/views/select-location/hooks/index.ts
@@ -1,3 +1,2 @@
-export * from './use-has-custom-lists';
 export * from './use-is-location-selector-expanded';
 export * from './use-is-location-selector-isolated';
