--- desktop/packages/mullvad-vpn/src/renderer/components/views/select-location/hooks/use-has-custom-lists.ts
@@ -1,9 +1,0 @@
-import { useSelectLocationViewContext } from '../SelectLocationViewContext';
-
-export function useHasCustomLists() {
-  const { customListLocations } = useSelectLocationViewContext();
-
-  const hasCustomLists = customListLocations.length > 0;
-
-  return hasCustomLists;
-}
